Splits Accessibility JS into two files
