+++ src/js/accessibility/modal-focus.js
@@ -1,51 +1,4 @@
-import { getClass } from './functions';
-
 ( ( d ) => {
-
-	/*
-	*
-	* TAB-ABLE SUB MENUS ( MAIN MENU )
-	*
-	*/
-
-	// Get all top-level anchors.
-	const childNavLinks = d.querySelectorAll( '.main-navigation ul > li > a' );
-
-	// If they exist.
-	if ( childNavLinks ) {
-
-		// Run through each of them.
-		childNavLinks.forEach( childNavLink => {
-
-			// Set them up with a focus eventListener.
-			childNavLink.addEventListener( 'focus', () => {
-
-				// Get the class generated by WP when an li has children.
-				const subMenuClass = getClass( 'menu-item-has-children', childNavLink.parentNode.classList );
-
-				// Hide all child-menus by default.
-				d.querySelectorAll( '.sub-menu' ).forEach( subMenu => {
-
-					// If the child-menu has no children in focus, hide it.
-					if ( subMenu.contains( document.activeElement ) === false ) {
-						subMenu.style.display = '';
-					}
-
-				} );
-
-				// If the parent li of the current anchor being focus has children.
-				if ( 'menu-item-has-children' === subMenuClass ) {
-
-					// Find the closest parent child-menu.
-					const siblingSubMenu = childNavLink.parentNode.querySelector( '.sub-menu' );
-					// Set it to display.
-
-					siblingSubMenu.style.display = 'block';
-				}
-
-			} );
-		} );
-	}
 
 	/*
 	*
@@ -64,7 +17,7 @@ import { getClass } from './functions';
 		if ( undefined !== modalTarget && event.which === 13 ) {
 
 			// Set the top close button inside the modal as a var.
-			const modalTopClose = document.querySelector( modalTarget + ' .modal-header .close' );
+			const modalTopClose = d.querySelector( modalTarget + ' .modal-header .close' );
 
 			// setTimeout so it doesn't fire immediately and get hijacked by browser.
 			setTimeout( () => {
@@ -84,7 +37,7 @@ import { getClass } from './functions';
 			setTimeout( () => {
 
 				// Focus back on the top close button.
-				document.querySelector( '#' + modalId + ' .modal-header .close' ).focus();
+				d.querySelector( '#' + modalId + ' .modal-header .close' ).focus();
 			}, 5 );
 		}
 
@@ -98,7 +51,7 @@ import { getClass } from './functions';
 			setTimeout( () => {
 
 				// Focus back on the top close button.
-				document.querySelector( '#' + modalId + ' .modal-footer button' ).focus();
+				d.querySelector( '#' + modalId + ' .modal-footer button' ).focus();
 			}, 5 );
 		}
