+++ src/js/theme.js
@@ -2,7 +2,9 @@ import './theme/navigation.js';
 import './theme/lazyload.js';
 import './theme/skip-link-focus-fix.js';
 import './theme/youtube-embed.js';
-import './theme/accessibility.js';
+
+import './accessibility/nested-navigation.js';
+import './accessibility/modal-focus.js';
 
 import './forms-tracking/set-cookies.js';
 import './forms-tracking/set-forms-fields.js';
